Implemented a DB + Worked on backend
First, I created an online database that uses MySQL. Added the user info in a json file `cloudDBinfo.json`, it's not a secure way of handling such data but it's all I can come up with for now.
Second, I worked on the authentication backend for the sellers role, created `loginSeller.php` and added a place holder `home.html` and `login_seller.html` for testing purposes.